Feature/TeamB-164: Make attendance record id auto-increment

+++ Backend/src/main/java/com/example/CheckInApp/repository/AttendanceRecordRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AttendanceRecordRepository extends JpaRepository<AttendanceRecord, Float> {
+public interface AttendanceRecordRepository extends JpaRepository<AttendanceRecord, Long> {
 
     Optional<AttendanceRecord> findByEvent_IdAndUser_Id(Long eventId, Long userId);
